Fixed the bug that two ValueSetTemplates associated with a same variable had incompatible types: one has complete array type while the other has incomplete array type.
Previously, a write-footprint was always saved before the state being
changed by the write operation.  The fix swaps the order of them.
This rules out the last case where a write-footprint is on an
uninitialized variable of an aggregate type.

Besides, to force ValueSetTemplates to be free of incomplete array
types (except for the heap type), I imposed a (maybe temporary)
restriction on the use of "capture" of read/write-footprints:

No variable involving sequence type sub-objects shall be captured!




git-svn-id: svn://vsl.cis.udel.edu/civl/trunk@5372 fb995dde-84ed-4084-dfe6-e5aef3e2452c